feat: 이미지 파일 업로드/로드 로직 구현, 양조장 회원가입 시 이미지 업로드하도록 변경

+++ src/main/java/com/example/monghyang/domain/brewery/main/entity/BreweryImage.java
@@ -28,16 +28,25 @@ public class BreweryImage {
     @Column(nullable = false)
     private Integer seq;
     @Column(nullable = false)
-    private Integer volume;
+    private Long volume;
 
-    private BreweryImage(Brewery brewery, UUID imageKey, Integer seq, Integer volume) {
+    private BreweryImage(Brewery brewery, UUID imageKey, Integer seq, Long volume) {
         this.brewery = brewery;
         this.imageKey = imageKey;
         this.seq = seq;
         this.volume = volume;
     }
 
-    public static BreweryImage breweryKeySeqVolume(Brewery brewery, UUID imageKey, Integer seq, Integer volume) {
+
+    /**
+     * BreweryImage 엔티티를 저장하고 저장된 엔티티를 반환합니다.
+     * @param brewery 양조장 Entity
+     * @param imageKey 이미지 업로드 후 반환된 UUID 값
+     * @param seq 해당 이미지의 순서 정보
+     * @param volume 해당 이미지의 용량 정보
+     * @return 저장 후 생성된 이미지 UUID key
+     */
+    public static BreweryImage breweryKeySeqVolume(Brewery brewery, UUID imageKey, Integer seq, Long volume) {
         return new BreweryImage(brewery, imageKey, seq, volume);
     }
